tfslite-sdk: add status callbacks for file upload
Accept optional callback functions on a FileUpload object
to provide status on transaction preparation, sending, and
waiting phases.

This support is implemented for both standard Rust as well as
WASM users.

As part of this work, make some changes in how we grab and drop
locked references to the state store. Reorganize some of the state
store access calls to make the code consistent.

Signed-off-by: William Katsak <bill@taekion.com>